+++ frontend/src/styles/HardwareSet.css
