--- frontend/src/utils/types.ts
@@ -1,7 +1,0 @@
-export interface HardwareSet {
-    id: number;
-    projectid: string;
-    name: string;
-    capacity: number;
-    available: number;
-}
